+++ PetSyncReact/src/componentes/Online.jsx
@@ -1,0 +1,13 @@
+import "../estilos/online.css";
+
+export default function Online({user}) {
+  return (
+    <li className="rightbarFriend">
+      <div className="rightbarProfileImgContainer">
+        <img className="rightbarProfileImg" src={user.profilePicture} alt="" />
+        <span className="rightbarOnline"></span>
+      </div>
+      <span className="rightbarUsername">{user.username}</span>
+    </li>
+  );
+}
