+++ PetSyncReact/src/componentes/Rightbar.jsx
@@ -1,0 +1,68 @@
+import "../estilos/rightbar.css";
+import { Users } from "../dummyData";
+import Online from "../componentes/Online";
+
+export default function Rightbar({ profile }) {
+  const HomeRightbar = () => {
+    return (
+      <>
+        <div className="birthdayContainer">
+          <img className="birthdayImg" src="assets/gift.png" alt="" />
+          <span className="birthdayText">
+            <b>Carlos Foster</b> y <b>Otros 3 amigos</b> Cumplen años hoy.
+          </span>
+        </div>
+      </>
+    );
+  };
+
+  const ProfileRightbar = () => {
+    return (
+      <>
+        <h4 className="rightbarTitle">Amigos Conectados</h4>
+
+        <h4 className="rightbarTitle">Online Friends</h4>
+        <ul className="rightbarFriendList">
+          {Users.map((u) => (
+            <Online key={u.id} user={u} />
+          ))}
+        </ul>
+
+        <div className="rightbarFollowings">
+          <div className="rightbarFollowing">
+            <img
+              src="assets/person/1.jpeg"
+              alt=""
+              className="rightbarFollowingImg"
+            />
+            <span className="rightbarFollowingName">John Carter</span>
+          </div>
+          <div className="rightbarFollowing">
+            <img
+              src="assets/person/2.jpeg"
+              alt=""
+              className="rightbarFollowingImg"
+            />
+            <span className="rightbarFollowingName">John Carter</span>
+          </div>
+          <div className="rightbarFollowing">
+            <img
+              src="assets/person/3.jpeg"
+              alt=""
+              className="rightbarFollowingImg"
+            />
+            <span className="rightbarFollowingName">John Carter</span>
+          </div>
+        </div>
+      </>
+    );
+  };
+  
+  return (
+    <div className="rightbar">
+      <div className="rightbarWrapper">
+        {profile ? <ProfileRightbar /> : <HomeRightbar />}
+      </div>
+    </div>
+  );
+}
