+++ PetSyncReact/src/main.jsx
@@ -8,9 +8,9 @@ import 'bootstrap-icons/font/bootstrap-icons.css';
 //import { MainFeed } from './componentes/MainFeed.jsx'
 import "../src/estilos/stylesheetMainFeed.css"
 
-import Sidebar from "../src/componentes/sidebar/Sidebar";
+import Sidebar from "./componentes/Sidebar";
 
-import Rightbar from "../src/components/rightbar/Rightbar";
+import Rightbar from "./componentes/Rightbar";
 //<MainFeed></MainFeed>
 createRoot(document.getElementById('root')).render(
   <StrictMode>
